requirements.txt
Flask==2.2.2
snowflake-connector-python==3.0.0
Werkzeug==2.2.2
